+++ src/common/request.ts
@@ -41,6 +41,7 @@ export type methods =
 
 export class Request {
   private readonly transport = Popsicle.createTransport({
+    maxBufferSize: Infinity, // Allow data of any size
     rejectUnauthorized: false, // Need to tell node to ignore bad ssl cert
     type: "text",
   })